+++ zheng-upms/zheng-upms-dao/src/main/java/com/ivy/zheng/upms/dao/mapper/UpmsPermissionMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface UpmsPermissionMapper {
-    long countByExample(UpmsPermissionExample example);
+    int countByExample(UpmsPermissionExample example);
 
     int deleteByExample(UpmsPermissionExample example);
